Implement uniform setting function
set_uniform is parameter overloaded; that is it accepts variable(s)
of arbitrary type and calls the appropriate glUniformXX function
accordingly.

This has the disadvantage of having to implement new overloads every time we
use new variable type/num in the main code, but it might boost ease of use,
code readability/maintainability and level of abstraction.