Modify Makefile to compile in both MinGW and OS X
For some reason the system environent variable PKG_CONFIG_PATH doesn't seem to
be passed through Makefile in mingw64-msys2.
We can manually set it by using PKG_CONFIG_PATH=$(PKG_CONFIG_PATH), but then
the backslashes in the variable (C:\msys64\mingw64\lib\pkgconfig) disappears
somehow and causes problem.
As a temporary solution sort out mingw64 specifically using an ifeq and
set the env manually. It is not beautiful, but it works.

Also, for the time being the windows binary produced by pkg-config --static
does NOT run in standalone mode and this should be remedied.