Change count type from int to size_t
In modern C, size_t is recommended over int for typing whatever
variables that measure size value of something. It is usually an
unsigned int underneath but the bit length is known to differ across
different platforms. Luckily Racket provides _size CType for consistent
interaction with these types, so we also follow the mainstream.