Implement read fail/lock-down preventing mechanism
When libplumloader hangs due to internal error, the parent racket
instance hangs "itself" and fall into eternal pit of un-SIGTERMable
madness of ones and zeros. To prevent this implement a basic
auto-bailing mechanism.

This is done by adding a boolean status value to the Mesh struct and
throwing exception in Racket when this status is false.