Eliminate global variable and use variable-passing
For this to work, Racket must be able to receive the Context object and
pass it to the main loop.

To do so, only declare the Context struct (which contains the C++ object
std::queue) and possibly declare only one field GLFWwindow* in the Racket
ffi code to make it oblivious of the C++ object.